requirements.txt
dnspython==2.4.2
MouseInfo==0.1.3
numpy==1.26.2
opencv-python==4.8.1.78
Pillow==10.1.0
pyaes==1.6.1
pyasn1==0.5.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pymongo==4.6.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dotenv==1.0.0
python3-xlib==0.15
pytweening==1.0.7
rsa==4.9
Telethon==1.33.0

gspread==5.12.2
psutil==5.9.7